Many changes in the feature2D style model and renderer Now it takes : -  expression for all style elements -  level rendering for symbolizer - complexe graphic representations
Added many examples

+++ feature2DStyle/src/main/java/org/orbisgis/style/utils/ColorHelper.java
@@ -75,7 +75,7 @@ public final class ColorHelper {
      * @param alpha [0;1]
      * @return new color with alpha channel
      */
-    public static Color getColorWithAlpha(Color c, double alpha) {
+    public static Color getColorWithAlpha(Color c, float alpha) {
         int a = (int) (MAX_RGB_VALUE * alpha);
 
         if (a < 0) {
